fixed some css
i fixed the menu css to no longer be in tables along with fixing the index page. i also added more contrast to the footer and navigation background and increased the text size. i still need to fix the about and contact pages, but i wanted to see if i was heading in the right direction.